chore: clear the workspace clippy backlog
`cargo clippy --workspace --all-targets -- -D warnings` now exits 0. It had 21
findings in `miasma-core`, and because that crate failed first the four
downstream ones were never reached -- a red gate hiding the next failure. CI
runs clippy advisory-only (no `-D warnings`), so main was green throughout and
none of this was visible there.

All changes are behaviour-preserving:

- `self_heal.rs`, `app.rs` x2: collapsible `if`. In the two `app.rs` cases the
  inner condition renders a button as a side effect; `&&` short-circuits, so
  the button is still not drawn when the outer condition is false.
- `coordinator.rs`: derive `Default` for `PublishOptions` instead of the hand
  written impl. `usize::default()` is 0, matching what it wrote.
- `node.rs`: match `OutgoingConnectionError { peer_id: Some(peer_id), .. }` in
  the arm pattern rather than re-testing inside. A `None` peer_id now falls to
  `_ => {}`, which is what the arm body already did with it.
- `reality.rs`: two `as u16` casts on values already `u16`.
- `store.rs` x3, `shadowsocks.rs` x8, `integration_test.rs`: needless borrows
  of arguments that already implement the required traits.
- `worker.rs`: drop `DirectedInboxItem::retention_secs`. It was carried over
  from `EnvelopeSummary` and never read -- both the inbox and outbox lists show
  the resulting `expires_at` as "Expires: ...", and there is no string or
  widget anywhere that displays the retention period itself.

818 workspace tests pass.

+++ crates/miasma-core/src/transport/reality.rs
@@ -743,8 +743,8 @@ mod tests {
         // SNI extension body
         let sni_bytes = sni.as_bytes();
         let sni_name_len = sni_bytes.len() as u16;
-        let sni_list_len = (3 + sni_name_len) as u16; // type(1) + len(2) + name
-        let sni_ext_len = (2 + sni_list_len) as u16; // list_len(2) + list
+        let sni_list_len = 3 + sni_name_len; // type(1) + len(2) + name
+        let sni_ext_len = 2 + sni_list_len; // list_len(2) + list
 
         let mut sni_ext = Vec::new();
         sni_ext.extend_from_slice(&[0x00, 0x00]); // ext_type: server_name
